Revert "Updated resources (Newtonsoft 12.0.2, NAudio 1.9.0) using NuGet"
This reverts commit df5d7607a27f027d009ce49dce5668c03d7dc0aa.

+++ EZBlocker2/MainForm.cs
@@ -86,7 +86,6 @@ namespace EZBlocker2
             movement.Exclude(typeof(Button));
             movement.Exclude(typeof(CheckBox));
             movement.Exclude(typeof(LinkLabel));
-
             movement.SetMovement(Controls);
 
             contextMenuStrip.Renderer = new CustomToolStripRenderer();
@@ -113,7 +112,10 @@ namespace EZBlocker2
                 try
                 {
                     string oldEZBlockerLog = ezBlockerLog + ".old";
-                    DeleteFile(oldEZBlockerLog);
+
+                    if (File.Exists(oldEZBlockerLog))
+                        File.Delete(oldEZBlockerLog);
+                    
                     File.Move(ezBlockerLog, oldEZBlockerLog);
                 }
                 catch { }
@@ -256,12 +258,12 @@ namespace EZBlocker2
             RegistryKey startupKey = Registry.CurrentUser.OpenSubKey(@"SOFTWARE\Microsoft\Windows\CurrentVersion\Run", true);
             foreach (string name in startupKey.GetValueNames())
             {
-                if ((name.ToLower().Contains("ezblocker") && !name.Equals(Program.ProductName)) || (name.Equals(Program.ProductName) && !checkBoxStartOnLogin.Checked))
+                if ((name.ToLower().Contains("ezblocker") && !name.Equals("EZBlocker 2")) || (name.Equals("EZBlocker 2") && !checkBoxStartOnLogin.Checked))
                     startupKey.DeleteValue(name); // Clean
             }
             if (checkBoxStartOnLogin.Checked)
             {
-                string name = Program.ProductName;
+                string name = "EZBlocker 2";
                 var value = startupKey.GetValue(name);
                 if (value == null || !value.Equals(ezBlockerFullExe))
                     startupKey.SetValue(name, ezBlockerFullExe);
@@ -373,15 +375,15 @@ namespace EZBlocker2
                 if (fix || write)
                 {
                     File.WriteAllLines(hostsFullFile, newLines);
-                    MessageBox.Show("Hosts file has been " + (fix ? "fixed" : "updated") + " successfully!", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    MessageBox.Show("Hosts file has been " + (fix ? "fixed" : "updated") + " successfully!", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             }
             catch
             {
                 if (!IsAdmin)
                 {
-                    string text = (loading ? "An error has been found while checking hosts file...\r\nTo solve this problem" : "To " + (checkBoxBlockAds.Checked ? "enable" : "disable") + " this option") + " administrator rights are required.\r\n\r\nDo you want to restart " + Program.ProductName + " in admin mode?";
-                    if (MessageBox.Show(text, Program.ProductName, MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes && RestartEZBlocker(true))
+                    string text = (loading ? "An error has been found while checking hosts file...\r\nTo solve this problem" : "To " + (checkBoxBlockAds.Checked ? "enable" : "disable") + " this option") + " administrator rights are required.\r\n\r\nDo you want to restart EZBlocker 2 in admin mode?";
+                    if (MessageBox.Show(text, "EZBlocker 2", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes && RestartEZBlocker(true))
                         return true;
                     else
                         return false;
@@ -397,7 +399,7 @@ namespace EZBlocker2
                     }
                     else
                     {
-                        MessageBox.Show("An error has been found while checking hosts file...\r\n" + Program.ProductName + " is unable to resolve this problem...", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                        MessageBox.Show("An error has been found while checking hosts file...\r\nEZBlocker 2 is unable to resolve this problem...", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Warning);
                         return false;
                     }
                 }
@@ -428,7 +430,7 @@ namespace EZBlocker2
 
             if (spotifyNotInstalled)
             {
-                MessageBox.Show("Spotify seems not to be installed on your system.\r\n" + Program.ProductName + " is useless for you...", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                MessageBox.Show("Spotify seems not to be installed on your system.\r\nEZBlocker 2 is useless for you...", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 KillEZBlocker();
             }
             else if (!IsSpotifyRunning())
@@ -462,10 +464,10 @@ namespace EZBlocker2
             catch
             {
                 if (IsAdmin)
-                    MessageBox.Show("Error while extracting dependencies!\r\n" + Program.ProductName + " will be closed immediately", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    MessageBox.Show("Error while extracting dependencies!\r\nEZBlocker 2 will be closed immediately", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 else
                 {
-                    if (MessageBox.Show("Error while extracting dependencies...\r\nDo you want to retry in admin mode?", Program.ProductName, MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
+                    if (MessageBox.Show("Error while extracting dependencies...\r\nDo you want to retry in admin mode?", "EZBlocker 2", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
                         RestartEZBlocker(true);
                 }
                 KillEZBlocker();
@@ -520,7 +522,7 @@ namespace EZBlocker2
 
                 ShowMessage("Hooking to Spotify...");
 
-                Spotify.WebAPI.OnNewStatus += Main_Status;
+                Spotify.WebAPI.NewStatus += Main_Status;
                 Spotify.WebAPI.RedirectUri = Uri.EscapeUriString(server.Prefix);
 
                 NameValueCollection data = HttpUtility.ParseQueryString(string.Empty);
@@ -543,20 +545,17 @@ namespace EZBlocker2
         {
             if (Spotify.WebAPI.APIToken != null)
             {
-                timerStatus.Enabled = false; // wait...
-
                 try
                 {
-                    if (processTmp != null)
+                    if (processTmp != null && !processTmp.HasExited)
                     {
-                        if (!processTmp.HasExited)
-                            processTmp.Kill();
-
+                        processTmp.Kill();
                         processTmp = null;
                     }
                 }
                 catch { }
                 
+                timerStatus.Enabled = false; // wait...
                 Spotify.WebAPI.GetStatus();
             }
         }
@@ -567,7 +566,7 @@ namespace EZBlocker2
             
             if (Spotify.WebAPI.Status.Retry_After == 0)
             {
-                timerStatus.Interval = 500; // 0.5s
+                timerStatus.Interval = 300;
 
                 if (IsSpotifyRunning())
                 {
@@ -600,7 +599,7 @@ namespace EZBlocker2
                                 ShowMessage("Spotify is in pause");
                         }
                         else
-                            ShowMessage("Spotify is in private session", "Disable private session to allow " + Program.ProductName + " to work");
+                            ShowMessage("Spotify is in private session", "Disable private session to allow EZBlocker 2 to work");
                     }
                     else
                         ShowMessage("Error: " + Spotify.WebAPI.Status.Error.Message, "Error: " + Spotify.WebAPI.Status.Error.Message);
@@ -609,7 +608,7 @@ namespace EZBlocker2
                 {
                     enable = false; // stop!
                     MinimizeEZBlocker();
-                    notifyIcon.ShowBalloonTip(3000, Program.ProductName, "Exiting from " + Program.ProductName + "...", ToolTipIcon.Info);
+                    notifyIcon.ShowBalloonTip(3000, "EZBlocker 2", "Exiting from EZBlocker 2...", ToolTipIcon.Info);
                     CloseEZBlocker(3000);
                 }
             }
@@ -624,7 +623,7 @@ namespace EZBlocker2
         {
             if (e.KeyCode == Keys.ControlKey)
             {
-                if (MessageBox.Show("Do you want to move " + Program.ProductName + " in the middle of the screen?", Program.ProductName, MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                if (MessageBox.Show("Do you want to move EZBlocker in the middle of the screen?", "EZBlocker 2", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                 {
                     Location = centerLocation;
                     SaveLocation();
@@ -652,7 +651,7 @@ namespace EZBlocker2
             else if (File.Exists(sndVol32FullExe))
                 StartProcess(sndVol32FullExe, "-m " + l.ToString());
             else
-                MessageBox.Show("Could not open system volume mixer.", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                MessageBox.Show("Could not open system volume mixer.", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Warning);
         }
 
         private void BtnSndVol_MouseLeave(object sender, EventArgs e) => btnSndVol.BackgroundImage = Properties.Resources.SndVol_Leave;
@@ -692,7 +691,7 @@ namespace EZBlocker2
 
                 if (BlockAds())
                 {
-                    MessageBox.Show("It could be necessary to restart Spotify or your computer for this changes to take effect.", Program.ProductName, MessageBoxButtons.OK, MessageBoxIcon.Information);
+                    MessageBox.Show("It could be necessary to restart Spotify or your computer for this changes to take effect.", "EZBlocker 2", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     Properties.Settings.Default.Save();
                 }
                 else
